Added section on output redirection
I reckon output redirection is pretty important to cover. Wrote a section on it underneath 'chaining commands'. Also mentioned the nohup command and some links to other shells (because I want to inspire freshers to be giant unix-hipsters like me and use fish) -- you can remove that stuff if it's just cluttering up the cheatsheet.

How much do we want to cover in this thing? What about the if-statement and for/while loops?